# Conflicts:
#	ses-app/ses-admin-dev/pom.xml

+++ ses-api/ses-api-mobile-b/src/main/java/com/redescooter/ses/api/mobile/b/vo/express/OrderResult.java
@@ -195,7 +195,7 @@ public class OrderResult extends GeneralResult {
 
 
     @ApiModelProperty(value="更新人")
-    private Integer updatedBy;
+    private Long updatedBy;
 
 
     @ApiModelProperty(value="更新时间")
